+++ src/tools/str/str_diff_t.hpp
@@ -9,7 +9,7 @@ namespace tools::str {
 
     // Struct to represent a single diff block
     typedef struct {
-        int bound[2];          // Line range where the difference occurs [start, end]
+        size_t bound[2];          // Line range where the difference occurs [start, end]
         vector<string> added;  // Lines added in s2
         vector<string> removed; // Lines removed from s1
     } str_diff_t;
